refactor(codeql): break 3 import cycles via neutral leaf modules + fix 2 mixed-returns
Clear the remaining 8 low CodeQL code-scanning alerts (6 py/cyclic-import +
2 py/mixed-returns; all note-level, zero security-severity).

CodeQL counts function-level/lazy imports in its static graph, so the existing
lazy-import workarounds did NOT clear the cyclic-import alerts. Break each cycle
by extracting the shared symbol the back-edge needs into a dependency-free leaf
module that both ends import:
  - lib/git_base.py — git primitives (GitError, run_git, main_repo_root,
    resolve_git_dirs, is_worktree) out of worktree_isolation
  - phase_quality/_findings_model.py — finding model + loaders out of _aggregates
  - shipwright-security/.../finding_classify.py — classify_finding out of aikido_client

The old homes re-export the moved names, so every external caller is unchanged.
Behavior is identical (full suites green: shared 3561, security 424, compliance
734; moved symbols verified to resolve to the same objects). worktree_isolation
shrinks below its bloat baseline.

The two py/mixed-returns are fixed by ending each test helper in an explicit
return/raise instead of a pytest.skip/fail fall-through (skip/fail semantics
preserved).

Run-ID: iterate-2026-06-28-codeql-import-cycles
