+++ .shipwright/agent_docs/build_dashboard.md
@@ -1,10 +1,11 @@
 # Project Activity Dashboard
-> Updated: 2026-06-28 06:17 UTC | Session: a9af39f8-f3c8-445b-a07a-86412ffe4704 | Run: iterate-2026-06-28-codeql-fixture-noise
+> Updated: 2026-06-28 07:10 UTC | Session: a9af39f8-f3c8-445b-a07a-86412ffe4704 | Run: iterate-2026-06-28-codeql-import-cycles
 
-## Recent Changes (213 iterations)
+## Recent Changes (214 iterations)
 
 | Type | Description | Tests | Commit | FRs | Date |
 |------|-------------|-------|--------|-----|------|
+| change | Break 3 CodeQL py/cyclic-import cycles via neutral leaf extraction + fix 2 py/mixed-returns | 4719/4719 |  | tooling | 2026-06-28 |
 | change | CodeQL fixture-noise cleanup: paths-ignore test fixtures + explicit string-concat refactor | 4720/4720 |  | infra | 2026-06-28 |
 | change | compliance PreToolUse Bash gates: robust uv run --no-project invocation + fail-open guard | 3560/3560 |  | tooling | 2026-06-27 |
 | change | CodeQL security hardening: tailor the query suite via codeql-config.yml; root-fix genuine findings (file modes to 0o600, two ReDoS regexes, a loop-capture bug, a rollback-CLI else-guard); remove 13 dead module globals. | 0/0 |  | tooling | 2026-06-27 |
@@ -220,7 +221,7 @@
 | change | post-adoption framework cleanup (Sub-1A through 1D) | 225/225 | 3db485b | FR-01.01, FR-01.02, FR-01.03 | 2026-05-02 |
 
 ## Test Status
-Last run: 2026-06-28 | Unit: 4720/4720 | Smoke: not_run | (iterate)
+Last run: 2026-06-28 | Unit: 4719/4719 | Smoke: not_run | (iterate)
 
 ## Pipeline
