+++ .shipwright/agent_docs/conventions.md
@@ -25,6 +25,7 @@ Python 3.11+ with uv as package manager. All scripts are invoked via uv run. Hoo
 
 > **One line per learning, no bold paragraphs** — always-loaded Layer-1 context. Format: `- (YYYY-MM-DD) <phase> — <one-line rule/gotcha>. → <run_id|ADR-NNN>`. **Budget ≤ 600 chars; if a learning won't fit one line it is a *decision* — write an ADR and leave a one-line pointer here.** Enforced repo-agnostically by the F11 verifier + `shared/scripts/tools/check_agent_doc_budget.py`; see `references/reflection.md`. Full rationale for compacted entries is in [`../planning/adr/_archive-agent-doc-updates.md`](../planning/adr/_archive-agent-doc-updates.md).
 
+- (2026-06-28) iterate — CodeQL `py/cyclic-import` counts function-level/lazy imports in its STATIC graph, so making an import lazy does NOT clear it (a `# noqa: PLC0415` lazy import was still flagged). Break a cycle by extracting the shared symbol the back-edge needs into a dependency-free LEAF module both ends import (re-export from the old home so callers are unchanged). Same shape clears `py/mixed-returns`: end the function in an explicit `return`/`raise`, never a `pytest.skip`/`fail` fall-through. → iterate-2026-06-28-codeql-import-cycles
 - (2026-06-28) iterate — CodeQL scanned a bundled JS test fixture → ~1000 `js/assignment-to-constant` alerts; the durable fix is a config `paths-ignore` for test fixtures (mirrors ruff `extend-exclude`), NOT per-alert dismissal (those reappear every scan). Clear an intentional `py/implicit-string-concatenation-in-list` by making it explicit `+` (ruff selects no ISC rule, so no gate conflict); prove the refactor byte-identical by comparing compiled `co_consts` before/after (the compiler folds both implicit and `+` literal concat). → iterate-2026-06-28-codeql-fixture-noise
 
 - (2026-06-28) iterate — a PreToolUse `Bash` gate runs on EVERY bash call, so its invocation must be robust and its body fail-OPEN, else a flaky/crashing check hard-blocks unrelated work. Two layers: `uv run --no-project` (skip the per-call project sync whose intermittent failure fail-closed as 'No stderr output') + route via `lib/hook_failopen.run_failopen` (exception → log `runtime/hook_errors.log`, return 0 ALLOW). The deliberate `return 2` soft-block survives (value, not exception). → iterate-2026-06-28-security-scan-hook-failopen
