--- .shipwright/agent_docs/iterates/iterate-2026-06-12-marketplace-python3-stub-probe.json
@@ -1,10 +1,0 @@
-{
-  "run_id": "iterate-2026-06-12-marketplace-python3-stub-probe",
-  "date": "2026-06-12T18:29:14.851731Z",
-  "type": "bug",
-  "complexity": "small",
-  "branch": "iterate/marketplace-python3-stub-probe",
-  "spec": null,
-  "tests_passed": true,
-  "adr": "iterate-2026-06-12-marketplace-python3-stub-probe"
-}
