+++ .shipwright/agent_docs/iterates/iterate-2026-06-28-codeql-import-cycles.json
@@ -1,0 +1,10 @@
+{
+  "run_id": "iterate-2026-06-28-codeql-import-cycles",
+  "date": "2026-06-28T07:10:23.529652Z",
+  "type": "change",
+  "complexity": "small",
+  "branch": "iterate/codeql-import-cycles",
+  "spec": null,
+  "tests_passed": true,
+  "adr": "iterate-2026-06-28-codeql-import-cycles"
+}
