+++ .shipwright/agent_docs/session_handoff.md
@@ -1,35 +1,34 @@
 ---
 canon_generated: true
-run_id: "iterate-2026-06-28-codeql-fixture-noise"
+run_id: "iterate-2026-06-28-codeql-import-cycles"
 phase: "iterate"
-reason: "iterate: CodeQL fixture-noise cleanup"
-timestamp: "2026-06-28T06:17:09.227517+00:00"
+reason: "iterate: break CodeQL import cycles + mixed-returns"
+timestamp: "2026-06-28T07:10:14.008342+00:00"
 ---
 
 # Session Handoff
 
-> Auto-generated 2026-06-28 06:17:09 UTC
+> Auto-generated 2026-06-28 07:10:14 UTC
 
 ## Session Info
 
 - **Session ID**: a9af39f8-f3c8-445b-a07a-86412ffe4704
-- **Timestamp**: 2026-06-28 06:17:09 UTC
-- **Reason**: iterate: CodeQL fixture-noise cleanup
+- **Timestamp**: 2026-06-28 07:10:14 UTC
+- **Reason**: iterate: break CodeQL import cycles + mixed-returns
 
 ## Last Iterate
 
-- **Run ID**: iterate-2026-06-28-security-scan-hook-failopen
-- **Date**: 2026-06-27T22:43:35.568860Z
+- **Run ID**: iterate-2026-06-28-codeql-fixture-noise
+- **Date**: 2026-06-28T06:17:40.642279Z
 - **Type**: change
-- **Complexity**: medium
-- **Branch**: iterate/2026-06-28-security-scan-hook-failopen
-- **ADR**: iterate-2026-06-28-security-scan-hook-failopen
+- **Complexity**: small
+- **Branch**: iterate/codeql-fixture-noise
+- **ADR**: iterate-2026-06-28-codeql-fixture-noise
 - **Tests passed**: True
-- **Spec**: .shipwright/planning/iterate/2026-06-28-security-scan-hook-failopen.md
 
 ## Current Iterate Progress
 
-- **Branch**: iterate/codeql-fixture-noise
+- **Branch**: iterate/codeql-import-cycles
 - **External Review Marker**: missing
 
 ### Mandatory replay on Resume
@@ -48,8 +47,8 @@ Before dispatching to the handoff's Remaining phase, run these if missing:
 
 ## Git State
 
-- **Branch**: iterate/codeql-fixture-noise
-- **Last Commit**: 01059128 fix(compliance): fail-open + robust invocation for PreToolUse Bash gates (#278)
+- **Branch**: iterate/codeql-import-cycles
+- **Last Commit**: 6f1f3507 chore(security): exclude test fixtures from CodeQL + make intentional string-concat explicit (#279)
 - **Uncommitted Changes**: Yes
 
 ## Config Files to Read
@@ -65,17 +64,17 @@ Before dispatching to the handoff's Remaining phase, run these if missing:
 
 | Event | Type | Source | Date |
 |-------|------|--------|------|
+| evt-bc8ebee5 | work_completed | iterate (Break 3 CodeQL py/cyclic-import cycles via neutral leaf extraction + fix 2 py/mixed-returns) | 2026-06-28 |
 | evt-5d34869b | work_completed | iterate (CodeQL fixture-noise cleanup: paths-ignore test fixtures + explicit string-concat refactor) | 2026-06-28 |
 | evt-d50b793d | work_completed | iterate (compliance PreToolUse Bash gates: robust uv run --no-project invocation + fail-open guard) | 2026-06-27 |
 | evt-e5afeb65 | work_completed | iterate (CodeQL security hardening: tailor the query suite via codeql-config.yml; root-fix genuine findings (file modes to 0o600, two ReDoS regexes, a loop-capture bug, a rollback-CLI else-guard); remove 13 dead module globals.) | 2026-06-27 |
 | evt-2dbacb5b | work_completed | iterate (Control Grade scorer (lib/control_grade.py, in Anlehnung an OpenSSF Scorecard) + Control Verdict block atop the dashboard (AR-01); latest-full-suite resolver kills the 0/0 headline in dashboard + test-evidence (AR-02); inline consistency-audit summary replaces the dead gitignored audit-report.md link (AR-03).) | 2026-06-27 |
-| evt-cda28075 | work_completed | iterate (Add shared/tests/test_trivyignore_register.py enforcing that every .trivyignore.yaml accepted-risk entry is scoped (paths|purls) + time-bounded (expired_at) + justified (statement); register optional (absent passes). Self-tested (rejects sloppy, accepts well-formed).) | 2026-06-22 |
 
 ## Recovery
 
 - **Pipeline**: 1 phases completed
-- **Total work events**: 213
-- **Last iterate**: change — CodeQL fixture-noise cleanup: paths-ignore test fixtures + explicit string-concat refactor (2026-06-28)
+- **Total work events**: 214
+- **Last iterate**: change — Break 3 CodeQL py/cyclic-import cycles via neutral leaf extraction + fix 2 py/mixed-returns (2026-06-28)
 - **Resume**: `/shipwright-iterate` for next change, or `/shipwright-run` for new pipeline
 
 ## Recent Decisions
