+++ .shipwright/agent_docs/triage_inbox.md
@@ -1,6 +1,6 @@
 # Triage Inbox
 
-> Auto-generated 2026-06-28T06:17:09.227517Z. Items waiting for triage decision.
+> Auto-generated 2026-06-28T07:10:14.008342Z. Items waiting for triage decision.
 > Promote via WebUI Triage tab (when v1b lands) or `shared/scripts/tools/triage_promote.py --id <id> --task-ref EXT:<ref>`.
 
 ## Status summary
