+++ .shipwright/compliance/change-history.md
@@ -1,7 +1,7 @@
 # Commit Change Log
 
-Generated: 2026-06-28T06:17:09.227517+00:00
-Total commits: 1043
+Generated: 2026-06-28T07:10:14.008342+00:00
+Total commits: 1044
 
 ## Commit Distribution
 
@@ -9,7 +9,7 @@ Total commits: 1043
 pie title Commit Types
     "fix" : 297
     "feat" : 275
-    "chore" : 228
+    "chore" : 229
     "docs" : 141
     "refactor" : 62
     "test" : 25
@@ -602,10 +602,11 @@ pie title Commit Types
 | 2026-03-20 | — | Task 02 — project templates (CLAUDE.md, agent_docs, CI) | c3a6d2f53bd3 |
 | 2026-03-20 | — | Task 01 — monorepo scaffolding + supabase-nextjs stack profile | 990a138a4690 |
 
-### Chores (chore) — 228 commits
+### Chores (chore) — 229 commits
 
 | Date | Scope | Description | Commit |
 |------|-------|-------------|--------|
+| 2026-06-28 | security | exclude test fixtures from CodeQL + make intentional string-concat explicit (#279) | 6f1f35076dfa |
 | 2026-06-22 | release | v0.29.1 (#271) | af7a3fc04fae |
 | 2026-06-17 | — | align root pyproject.toml to 0.29.0 + de-PII a source comment (#262) | 6c0c1ea626c4 |
 | 2026-06-17 | release | v0.29.0 (#261) | 392152b988ec |
@@ -1112,7 +1113,7 @@ pie title Commit Types
 
 | Metric | Value |
 |--------|-------|
-| Total commits | 1043 |
+| Total commits | 1044 |
 | AI-assisted commits | 0 |
-| Human-authored commits | 1043 |
+| Human-authored commits | 1044 |
