+++ .shipwright/compliance/sbom.md
@@ -1,6 +1,6 @@
 # Software Bill of Materials (SBOM)
 
-Generated: 2026-06-28T06:17:09.227517+00:00
+Generated: 2026-06-28T07:10:14.008342+00:00
 
 ## Summary
