+++ CHANGELOG-unreleased.d/Fixed/iterate-2026-06-28-codeql-import-cycles_001.md
@@ -1,0 +1,1 @@
+Cleared two py/mixed-returns code-scanning alerts by ending the affected test helpers in an explicit return/raise instead of a pytest.skip/fail fall-through
