+++ shipwright_events.jsonl
@@ -216,3 +216,4 @@
 {"intent":"change","spec_impact":"none","change_type":"tooling","none_reason":"Framework-internal security hardening remediating CodeQL findings, plus a monorepo-only CodeQL config tailoring; no tracked FR or planning spec governs these scripts or the CI config.","description":"CodeQL security hardening: tailor the query suite via codeql-config.yml; root-fix genuine findings (file modes to 0o600, two ReDoS regexes, a loop-capture bug, a rollback-CLI else-guard); remove 13 dead module globals.","v":1,"id":"evt-e5afeb65","ts":"2026-06-27T21:10:50.533812+00:00","type":"work_completed","source":"iterate","commit":"","adr_id":"iterate-2026-06-27-codeql-security-hardening","session":"12c9c16a-2f03-47d2-9d1b-612113aa9379"}
 {"intent":"change","description":"compliance PreToolUse Bash gates: robust uv run --no-project invocation + fail-open guard","spec_impact":"none","spec_impact_justification":"Behavior-preserving for the gate decision (the deploy/commit soft-block + override are unchanged and regression-tested); only adds invocation robustness and a fail-open crash path. No product FR.","change_type":"tooling","none_reason":"Framework hook robustness — affects only the compliance PreToolUse Bash gate machinery; no product feature/FR.","changed_files":["plugins/shipwright-compliance/hooks/hooks.json","plugins/shipwright-compliance/scripts/hooks/check_security_scan.py","plugins/shipwright-compliance/scripts/hooks/check_rtm_coverage.py","plugins/shipwright-compliance/scripts/lib/hook_failopen.py","plugins/shipwright-compliance/tests/test_hook_failopen.py","integration-tests/test_compliance_hook_failopen.py","docs/hooks-and-pipeline.md"],"tests":{"passed":3560,"total":3560,"e2e_run":false},"v":1,"id":"evt-d50b793d","ts":"2026-06-27T22:43:07.150922+00:00","type":"work_completed","source":"iterate","commit":"","adr_id":"iterate-2026-06-28-security-scan-hook-failopen","session":"a0dc1dca-ceaf-4731-b3db-1d33dd94cc32"}
 {"intent":"change","description":"CodeQL fixture-noise cleanup: paths-ignore test fixtures + explicit string-concat refactor","spec_impact":"none","spec_impact_justification":"Behavior-preserving: CI CodeQL scan-config change plus byte-identical string-concat refactors; no application/spec behavior changes (verified by a compile-constant equivalence gate).","change_type":"infra","none_reason":"CI security-scanning config plus a code-quality refactor; touches no FR or spec.","tests":{"passed":4720,"total":4720,"e2e_run":false},"changed_files":[".github/codeql/codeql-config.yml","shared/tests/test_codeql_config_query_filters.py","plugins/shipwright-compliance/scripts/lib/_control_block.py","plugins/shipwright-security/scripts/tools/generate_security_report.py","plugins/shipwright-adopt/tests/test_data_preservation_realistic.py","shared/tests/test_installer_shell_scripts.py"],"v":1,"id":"evt-5d34869b","ts":"2026-06-28T06:17:09.227517+00:00","type":"work_completed","source":"iterate","commit":"","adr_id":"iterate-2026-06-28-codeql-fixture-noise","session":"a9af39f8-f3c8-445b-a07a-86412ffe4704"}
+{"intent":"change","description":"Break 3 CodeQL py/cyclic-import cycles via neutral leaf extraction + fix 2 py/mixed-returns","spec_impact":"none","spec_impact_justification":"Behavior-preserving refactor: moved shared symbols into dependency-free leaf modules (re-exported) and reshaped two test helpers' return paths. No application/spec behavior changes (full suites green + import-identity verified).","change_type":"tooling","none_reason":"Internal import-cycle refactor + test-helper return-shape fixes to clear CodeQL code-quality alerts; touches no FR or spec.","tests":{"passed":4719,"total":4719,"e2e_run":false},"changed_files":["shared/scripts/lib/git_base.py","shared/scripts/lib/worktree_isolation.py","shared/scripts/lib/repo_root.py","shared/scripts/lib/phase_quality/_findings_model.py","shared/scripts/lib/phase_quality/_aggregates.py","shared/scripts/lib/phase_quality/_dashboard_render.py","plugins/shipwright-security/scripts/lib/finding_classify.py","plugins/shipwright-security/scripts/lib/scanner_backend.py","plugins/shipwright-security/scripts/lib/aikido_client.py","shared/tests/test_artifact_drift_edge_cases.py","shared/tests/test_security_critical_gate.py"],"v":1,"id":"evt-bc8ebee5","ts":"2026-06-28T07:10:14.008342+00:00","type":"work_completed","source":"iterate","commit":"","adr_id":"iterate-2026-06-28-codeql-import-cycles","session":"a9af39f8-f3c8-445b-a07a-86412ffe4704"}
